Made simpler implementation of subnet.NewSummary
